refactor/#9: error 패키지 경로 이동 및 온보딩 에러 코드 추가
